+++ Common/Database/BaseDatabaseData.cs
@@ -6,6 +6,6 @@ namespace EggLink.DanhengServer.Database
     public abstract class BaseDatabaseData
     {
         [SugarColumn(IsPrimaryKey = true)]
-        public long Uid { get; set; }
+        public int Uid { get; set; }
     }
 }
